Add Swiper features
