+++ navigation/AppStack.js
@@ -4,7 +4,8 @@ import { NavigationContainer } from "@react-navigation/native";
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 import { createStackNavigator } from "@react-navigation/stack";
 
-import HomeScreen from './../screens/HomeScreen';
+import CustomerScreen from '../screens/CustomerScreen';
+import HomeScreen from '../screens/HomeScreen';
 import ProfileScreen from "./../screens/ProfileScreen";
 import SellerScreen from "./../screens/SellerScreen";
 import AddProductScreen from "./../screens/AddProductScreen";
@@ -37,10 +38,11 @@ const AppStack = () => {
   return (
     <Tab.Navigator screenOptions={{ headerShown: false, tabBarStyle: { display: 'none' } }}>
       <Tab.Screen name="Home" component={HomeScreen} />
+      <Tab.Screen name="CustomerScreen" component={CustomerScreen} />
+      <Tab.Screen name="SellerScreen" component={SellerScreen} />
       <Tab.Screen name="ChatScreen" component={MessageStack} />
       <Tab.Screen name="Profile" component={ProfileScreen} />
       <Tab.Screen name="EditProfile" component={EditProfileScreen} />
-      <Tab.Screen name="Seller" component={SellerScreen} />
       <Tab.Screen name="AddProduct" component={AddProductScreen} />
       <Tab.Screen name="EditProduct" component={EditProductScreen} />
       <Tab.Screen name="DetailsProduct" component={ProductDetailsScreen} />
