+++ package.json
@@ -28,14 +28,15 @@
     "react-native-gifted-chat": "^2.1.0",
     "react-native-keyboard-aware-scroll-view": "^0.9.5",
     "react-native-paper": "^5.7.2",
+    "react-native-reanimated": "~2.14.4",
     "react-native-safe-area-context": "^4.5.0",
     "react-native-screens": "~3.20.0",
     "react-native-star-rating-widget": "^1.7.1",
     "react-native-svg": "13.4.0",
+    "react-native-swiper": "^1.6.0",
     "react-native-vector-icons": "^9.2.0",
     "react-native-web": "~0.18.11",
-    "styled-components": "^6.0.0-rc.1",
-    "react-native-reanimated": "~2.14.4"
+    "styled-components": "^6.0.0-rc.1"
   },
   "resolutions": {
     "styled-components": "^5"
