+++ screens/CustomerScreen.js
@@ -1,0 +1,202 @@
+import { View, Text, StyleSheet, ScrollView, TextInput, TouchableOpacity } from 'react-native';
+import { SafeAreaView } from 'react-native-safe-area-context';
+import AsyncStorage from '@react-native-async-storage/async-storage';
+import { COLORS, SIZES } from '../constants/theme';
+import HeaderTabs from '../components/HeaderTabs';
+import CardItem from '../components/CardItem';
+import { Divider } from 'react-native-elements';
+import BottomTabs from '../components/BottomTabs';
+import { useState } from 'react';
+import React from 'react';
+import { useFocusEffect } from '@react-navigation/native';
+import axios from 'axios';
+import Ioniocons from "react-native-vector-icons/Ionicons";
+import { config } from '../config';
+import { useNavigation, useIsFocused } from '@react-navigation/native';
+
+const CustomerScreen = () => {
+  const [productos, setProductos] = useState();
+  const [filteredProducts, setFilteredProducts] = useState(undefined)
+  const [doProductsRequest, setDoProductsRequest] = useState(false)
+  const navigation = useNavigation();
+  const isFocused = useIsFocused()
+  const searchProduct = (text) => {
+    console.log(text);
+    if(text === ''){
+      setFilteredProducts(undefined)
+      setDoProductsRequest(!doProductsRequest)
+      return
+    }
+
+    const arrayFilteredProducts = productos.map(usuario => {
+      return {...usuario,
+        productos: usuario.productos.filter(producto =>{
+        const productName = producto.nombre.toLowerCase()
+        const searchText = text.toLowerCase()
+        return productName.startsWith(searchText)
+      })}
+    })
+
+    setFilteredProducts(arrayFilteredProducts)
+  }
+
+  useFocusEffect(
+    React.useCallback(() => {
+      axios.get(config.apiUrl + '/api/productos').then((req) => {
+        const allProductsData = req.data
+        setProductos(allProductsData)
+      }).catch(e => {setProductos(null)});
+      console.log(isFocused)
+    }, [doProductsRequest])
+  );
+
+  React.useEffect(() => {
+    const fetchData = async () => {
+      try {
+        const response = await axios.get(config.apiUrl + '/api/productos');
+        const allProductsData = response.data;
+        setProductos(allProductsData);
+      } catch (error) {
+        setProductos(null);
+      }
+    };
+  
+    fetchData();
+  }, [doProductsRequest]);
+  return (
+    <>
+      <View
+        style={{
+          backgroundColor: "white",
+        }}>
+
+        <View style={styles.searchContainer}>
+          <View style={styles.searchWrapper}>
+            <TextInput
+              onChangeText={(text) =>{
+                searchProduct(text)
+              }}
+              style={styles.searchInput}
+              placeholder='¿Que estás buscando?'
+            />
+          </View>
+          <TouchableOpacity
+            style={styles.btnIcon}
+          >
+            <Ioniocons name="search" size={25} color={COLORS.white} />
+          </TouchableOpacity>
+        </View>
+      </View>
+      <ScrollView showsVerticalScrollIndicator={false} >
+        <View style={{
+          flex: 1,
+          paddingHorizontal: 15
+        }}>
+
+          {filteredProducts === undefined 
+          ? productos ? productos.map(usuario => {
+            const {id: idUsuario, productos, nombreCompleto } = usuario
+            const nombreUsuario = nombreCompleto?.split(' ')[0]  + " " + nombreCompleto?.split(' ')[2]
+            return (productos.map(producto => {
+              return (<CardItem
+              isActive={true}
+              nombreProducto={producto?.nombre}
+              descripcion={producto.descripcion}
+              puntaje={producto.puntaje}
+              precio={producto.precio}
+              imgUrl={producto.imgUrl}
+              categoria={producto.categoria[0]?.nombre}
+              nombreMarca = {usuario.nombreMarca}
+              nombreUsuario={nombreUsuario}
+              idUsuario={idUsuario}
+              key={producto.id}
+              onPress={() =>
+                navigation.navigate('DetailsProduct', {
+                  telefono: usuario.telefonos[0],
+                  nombreMarca: usuario.nombreMarca,
+                  nombreProducto: producto.nombre,
+                  descripcion: producto.descripcion,
+                  puntaje: producto.puntaje,
+                  precio: producto.precio,
+                  imgUrl: producto.imgUrl,
+                  
+                  nombreUsuario: nombreUsuario,
+                  idUsuario: idUsuario,
+                })}
+              />)
+            }))
+          }) : <Text>No hay productos para mostrar</Text> : filteredProducts && filteredProducts.map(usuario => {
+            const { nombreCompleto, id: idUsuario, productos } = usuario
+            const nombreUsuario = nombreCompleto?.split(' ')[0] + nombreCompleto?.split(' ')[2]
+            return (productos.map(producto => {
+              return (<CardItem
+              isActive={true}
+              nombreProducto={producto?.nombre}
+              descripcion={producto.descripcion}
+              puntaje={producto.puntaje}
+              precio={producto.precio}
+              imgUrl={producto.imgUrl}
+              categoria={producto.categoria[0]?.nombre}
+              nombreMarca = {usuario.nombreMarca}
+              nombreUsuario={nombreUsuario}
+              idUsuario={idUsuario}
+              key={producto.id}
+              onPress={() =>
+                navigation.navigate('DetailsProduct', {
+                  telefono: usuario.telefonos[0],
+                  nombreMarca: usuario.nombreMarca,
+                  nombreProducto: producto.nombre,
+                  descripcion: producto.descripcion,
+                  puntaje: producto.puntaje,
+                  precio: producto.precio,
+                  imgUrl: producto.imgUrl,
+                  
+                  nombreUsuario: nombreUsuario,
+                  idUsuario: idUsuario,
+                })}
+              />)
+            }))
+          })}
+        </View>
+      </ScrollView>
+    </>
+  );
+}
+
+const styles = StyleSheet.create({
+  searchContainer: {
+    justifyContent: "center",
+    alignItems: "center",
+    flexDirection: "row",
+    marginTop: SIZES.large,
+    height: 50,
+    paddingHorizontal: 15
+  },
+  searchWrapper: {
+    flex: 1,
+    backgroundColor: 'red',
+    marginRight: SIZES.small,
+    justifyContent: "center",
+    alignItems: "center",
+    borderRadius: SIZES.medium,
+    height: "100%",
+  },
+  searchInput: {
+    width: "100%",
+    height: "100%",
+    paddingHorizontal: SIZES.medium,
+    padding: 1,
+    backgroundColor: 'white'
+  },
+  btnIcon: {
+    backgroundColor: COLORS.primary,
+    borderRadius: 30,
+    height: 45,
+    width: 45,
+    justifyContent: "center",
+    alignItems: "center",
+  }
+});
+
+
+export default CustomerScreen;
