+++ screens/HomeScreen.js
@@ -1,198 +1,55 @@
-import { View, Text, StyleSheet, ScrollView, TextInput, TouchableOpacity } from 'react-native';
-import { SafeAreaView } from 'react-native-safe-area-context';
-import AsyncStorage from '@react-native-async-storage/async-storage';
-import { COLORS, SIZES } from './../constants/theme';
-import HeaderTabs from '../components/HeaderTabs';
-import CardItem from '../components/CardItem';
-import { Divider } from 'react-native-elements';
-import BottomTabs from '../components/BottomTabs';
-import { useState } from 'react';
 import React from 'react';
-import { useFocusEffect } from '@react-navigation/native';
-import axios from 'axios';
-import Ioniocons from "react-native-vector-icons/Ionicons";
-import { config } from '../config';
-import { useNavigation } from '@react-navigation/native';
+import { View, Text, StyleSheet } from 'react-native';
+import { Divider } from 'react-native-elements'
+import Swiper from 'react-native-swiper';
+import CustomerScreen from './CustomerScreen';
+import SellerScreen from './SellerScreen';
+import BottomTabs from '../components/BottomTabs';
+import HeaderTabs from '../components/HeaderTabs';
+import { SafeAreaView } from 'react-native-safe-area-context';
+import { COLORS, SIZES } from '../constants/theme';
+import { useState } from 'react';
+
 
 const HomeScreen = () => {
-  const [productos, setProductos] = useState();
-  const [filteredProducts, setFilteredProducts] = useState()
-  const [doProductsRequest, setDoProductsRequest] = useState(false)
-  const navigation = useNavigation();
 
-  const searchProduct = (text) => {
-    console.log(text);
-    if(text === ''){
-      setFilteredProducts(undefined)
-      setDoProductsRequest(!doProductsRequest)
-      return
-    }
+  const [activeTab, setActiveTab] = useState(0);
 
-    const arrayFilteredProducts = productos.map(usuario => {
-      return {...usuario,
-        productos: usuario.productos.filter(producto =>{
-        const productName = producto.nombre.toLowerCase()
-        const searchText = text.toLowerCase()
-        return productName.startsWith(searchText)
-      })}
-    })
-
-    setFilteredProducts(arrayFilteredProducts)
-  }
-
-  useFocusEffect(
-    React.useCallback(() => {
-      axios.get(config.apiUrl + '/api/productos').then((req) => {
-        const allProductsData = req.data
-        setProductos(allProductsData)
-      }).catch(e => {setProductos(null)});
-    }, [doProductsRequest])
-  );
+  const handleSwipe = (index) => {
+    setActiveTab(index)
+   
+  };
 
   return (
-    <SafeAreaView
-      style={{
+
+    <SafeAreaView style={{
         backgroundColor: COLORS.lightWhite,
         flex: 1,
-
       }}>
-      <View
-        style={{
-          backgroundColor: "white",
-          padding: 15,
-        }}>
-        <HeaderTabs />
-        <View style={styles.searchContainer}>
-          <View style={styles.searchWrapper}>
-            <TextInput
-              onChangeText={(text) =>{
-                searchProduct(text)
-              }}
-              style={styles.searchInput}
-              placeholder='¿Que estás buscando?'
-            />
-          </View>
-          <TouchableOpacity
-            style={styles.btnIcon}
-          >
-            <Ioniocons name="search" size={25} color={COLORS.white} />
-          </TouchableOpacity>
+        <HeaderTabs isActive={activeTab} />
+        <Swiper  onIndexChanged={handleSwipe} loop={false} showsPagination={false}>
+            <CustomerScreen />
+            <SellerScreen />
+        </Swiper>
+        <Divider width={1} />
+        <View style={{ backgroundColor: "white" }}>
+            <BottomTabs />
         </View>
-      </View>
-      <ScrollView showsVerticalScrollIndicator={false} >
-        <View style={{
-          flex: 1,
-          padding: 15,
-        }}>
-
-          {filteredProducts === undefined 
-          ? productos ? productos.map(usuario => {
-            const { nombres, apellidoPaterno, id: idUsuario, productos } = usuario
-            const nombre = nombres.split(' ')[0]
-            const nombreUsuario = nombre + ' ' + apellidoPaterno
-            return (productos.map(producto => {
-              return (<CardItem
-              isActive={true}
-              nombreProducto={producto?.nombre}
-              descripcion={producto.descripcion}
-              puntaje={producto.puntaje}
-              precio={producto.precio}
-              imgUrl={producto.imgUrl}
-              categoria={producto.categoria[0]?.nombre}
-              nombreMarca = {usuario.nombreMarca}
-              nombreUsuario={nombreUsuario}
-              idUsuario={idUsuario}
-              key={producto.id}
-              onPress={() =>
-                navigation.navigate('DetailsProduct', {
-                  telefono: usuario.telefonos[0],
-                  nombreMarca: usuario.nombreMarca,
-                  nombreProducto: producto.nombre,
-                  descripcion: producto.descripcion,
-                  puntaje: producto.puntaje,
-                  precio: producto.precio,
-                  imgUrl: producto.imgUrl,
-                  
-                  nombreUsuario: nombreUsuario,
-                  idUsuario: idUsuario,
-                })}
-              />)
-            }))
-          }) : <Text>No hay productos para mostrar</Text> : filteredProducts && filteredProducts.map(usuario => {
-            const { nombres, apellidoPaterno, id: idUsuario, productos } = usuario
-            const nombre = nombres.split(' ')[0]
-            const nombreUsuario = nombre + ' ' + apellidoPaterno
-            return (productos.map(producto => {
-              return (<CardItem
-              isActive={true}
-              nombreProducto={producto?.nombre}
-              descripcion={producto.descripcion}
-              puntaje={producto.puntaje}
-              precio={producto.precio}
-              imgUrl={producto.imgUrl}
-              categoria={producto.categoria[0]?.nombre}
-              nombreMarca = {usuario.nombreMarca}
-              nombreUsuario={nombreUsuario}
-              idUsuario={idUsuario}
-              key={producto.id}
-              onPress={() =>
-                navigation.navigate('DetailsProduct', {
-                  telefono: usuario.telefonos[0],
-                  nombreMarca: usuario.nombreMarca,
-                  nombreProducto: producto.nombre,
-                  descripcion: producto.descripcion,
-                  puntaje: producto.puntaje,
-                  precio: producto.precio,
-                  imgUrl: producto.imgUrl,
-                  
-                  nombreUsuario: nombreUsuario,
-                  idUsuario: idUsuario,
-                })}
-              />)
-            }))
-          })}
-        </View>
-      </ScrollView>
-      <Divider width={1} />
-      <View style={{ backgroundColor: "white" }}>
-        <BottomTabs />
-      </View>
-
     </SafeAreaView>
+
   );
-}
+};
 
 const styles = StyleSheet.create({
-  searchContainer: {
-    justifyContent: "center",
-    alignItems: "center",
-    flexDirection: "row",
-    marginTop: SIZES.large,
-    height: 50,
-  },
-  searchWrapper: {
+  screenContainer: {
     flex: 1,
-    backgroundColor: COLORS.white,
-    marginRight: SIZES.small,
-    justifyContent: "center",
-    alignItems: "center",
-    borderRadius: SIZES.medium,
-    height: "100%",
+    justifyContent: 'center',
+    alignItems: 'center',
   },
-  searchInput: {
-    width: "100%",
-    height: "100%",
-    paddingHorizontal: SIZES.medium,
+  screenText: {
+    fontSize: 24,
+    fontWeight: 'bold',
   },
-  btnIcon: {
-    backgroundColor: COLORS.primary,
-    borderRadius: 30,
-    height: 45,
-    width: 45,
-    justifyContent: "center",
-    alignItems: "center",
-  }
 });
-
 
 export default HomeScreen;
